kadaifinish

+++ app/Http/Controllers/TasksController.php
@@ -10,53 +10,55 @@ use App\Task;
 
 class TasksController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
+   
+   public function index()
     {
-        $tasks = Task::all();
+        if (\Auth::check()){
+            $user = \Auth::user();
+            $tasks = $user->tasks;
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
+            return view('tasks.index', [
+                'user' => $user,
+                'tasks' => $tasks
         ]);
+        } else {
+            return view('welcome');
+        }
+    }
+   
+   public function create()
+    {
+        if (\Auth::check()) {
+             $task=new Task;
+        
+             return view('tasks.create',[
+            'task'=>$task,
+            ]);
+        } else {
+            return view('welcome');
+        }
+       
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        $task = new Task;
-
-        return view('tasks.create', [
-            'task' => $task,
-        ]);
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+   
     public function store(Request $request)
     {
-        $this->validate($request, [
+         if (\Auth::check()){
+             $this->validate($request, [
+            'status' => 'required|max:10',   
             'content' => 'required|max:191',
-            'status' => 'required|max:10',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
+        $task=new Task;
+        $task->status=$request->status;
+        $task->content=$request->content;
+        $task->user_id=\Auth::user()->id;
         $task->save();
-
+        
         return redirect('/');
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -67,11 +69,20 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-       $task = Task::find($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+            if (\Auth::check()) {
+            $task=Task::find($id);
+            if ($task->user_id == \Auth::user()->id) {
+                
+                return view('tasks.show',[
+                    'task'=>$task,
+                ]);
+            } else {
+                return redirect('/');
+            }
+            
+             } else {
+                return redirect('welcome');
+             }
     }
 
     /**
@@ -82,11 +93,22 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::find($id);
-
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        
+         if (\Auth::check()) {
+            $task=Task::find($id);
+            if ($task->user_id == \Auth::user()->id) {
+                
+                return view('tasks.edit',[
+                    'task'=>$task,
+                ]);
+            } else {
+                return redirect('/');
+            }
+            
+             } else {
+                return redirect('welcome');
+             }
+        
     }
 
     /**
@@ -99,16 +121,15 @@ class TasksController extends Controller
     public function update(Request $request, $id)
     {
          $this->validate($request, [
+            'status' => 'required|max:10',   // add
             'content' => 'required|max:191',
-            'status' => 'required|max:10',
         ]);
         
-        
-        $task = Task::find($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
+        $task=Task::find($id);
+        $task->status=$request->status;
+        $task->content=$request->content;
         $task->save();
-
+        
         return redirect('/');
     }
 
@@ -118,11 +139,20 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+public function destroy($id)
     {
-       $task = Task::find($id);
-        $task->delete();
-
-        return redirect('/');
+        
+        if (\Auth::check()) {
+            $task=Task::find($id);
+            if ($task->user_id == \Auth::user()->id) {
+                
+                $task->delete();
+            }
+            return redirect('/');
+            
+        } else {
+            return view('welcome');
+        }
+      
     }
 }
